Split out builtin/forms.py into a module
In particular we separate:

* top-level forms in $OutputForms,
* top-level forms not in $OutputForms
* form variables $OutputForms, $PrintForms
* base class code

This is the non-experimental part of #588
