+++ mathics/builtin/forms/__init__.py
@@ -1,0 +1,11 @@
+"""
+Forms of Input and Output
+
+A <i>Form</i> format specifies the way Mathics Expression input is read or output written.
+
+The variable <url>:$OutputForms':
+http://localhost:8000/doc/reference-of-built-in-symbols/forms-of-input-and-output/form-variables/$outputforms/</url> has a list of Forms defined.
+
+See also <url>:WMA link:
+https://reference.wolfram.com/language/tutorial/TextualInputAndOutput.html#12368</url>.
+"""
