+++ mathics/core/systemsymbols.py
@@ -71,6 +71,7 @@ SymbolFactorial = Symbol("System`Factorial")
 SymbolFailed = Symbol("System`$Failed")
 SymbolFloor = Symbol("System`Floor")
 SymbolFormat = Symbol("System`Format")
+SymbolFractionBox = Symbol("System`FractionBox")
 SymbolFullForm = Symbol("System`FullForm")
 SymbolFunction = Symbol("System`Function")
 SymbolGamma = Symbol("System`Gamma")
@@ -120,6 +121,7 @@ SymbolNeeds = Symbol("System`Needs")
 SymbolNone = Symbol("System`None")
 SymbolNorm = Symbol("System`Norm")
 SymbolNot = Symbol("System`Not")
+SymbolNumberForm = Symbol("System`NumberForm")
 SymbolNumberQ = Symbol("System`NumberQ")
 SymbolNumericQ = Symbol("System`NumericQ")
 SymbolO = Symbol("System`O")
@@ -131,6 +133,7 @@ SymbolOutputForm = Symbol("System`OutputForm")
 SymbolOverflow = Symbol("System`Overflow")
 SymbolPackages = Symbol("System`$Packages")
 SymbolPattern = Symbol("System`Pattern")
+SymbolPower = Symbol("System`Power")
 SymbolPi = Symbol("System`Pi")
 SymbolPiecewise = Symbol("System`Piecewise")
 SymbolPoint = Symbol("System`Point")
@@ -158,12 +161,16 @@ SymbolSign = Symbol("System`Sign")
 SymbolSimplify = Symbol("System`Simplify")
 SymbolSin = Symbol("System`Sin")
 SymbolSlot = Symbol("System`Slot")
+SymbolSqrt = Symbol("System'Sqrt")
+SymbolSqrtBox = Symbol("System`SqrtBox")
 SymbolStandardForm = Symbol("System`StandardForm")
 SymbolStringForm = Symbol("System`StringForm")
 SymbolStringQ = Symbol("System`StringQ")
 SymbolStyle = Symbol("System`Style")
 SymbolSubsetQ = Symbol("System`SubsetQ")
 SymbolSubtract = Symbol("System`Subtract")
+SymbolSubscriptBox = Symbol("System`SubscriptBox")
+SymbolSubsuperscriptBox = Symbol("System`SubsuperscriptBox")
 SymbolSuperscriptBox = Symbol("System`SuperscriptBox")
 SymbolTable = Symbol("System`Table")
 SymbolTeXForm = Symbol("System`TeXForm")
